model: move code to handle model.Cmd -> exec.Cmd conversion to new package
